refactor: move moveTableRow and moveTableColumn to separate directory
- Create new packages/extensions/src/table/table-commands/ directory
- Move moveTableRow function and MoveTableRowOptions to move-table-row.ts
- Move moveTableColumn function and MoveTableColumnOptions to move-table-column.ts
- Update imports in table-commands.ts to reference new files
- Update table/index.ts to export moved functions directly from new locations
