+++ packages/extensions/src/table/index.ts
@@ -6,21 +6,25 @@ export {
   defineTableCommands,
   exitTable,
   insertTable,
-  moveTableColumn,
-  moveTableRow,
   selectTable,
   selectTableCell,
   selectTableColumn,
   selectTableRow,
   type InsertTableOptions,
-  type MoveTableColumnOptions,
-  type MoveTableRowOptions,
   type SelectTableCellOptions,
   type SelectTableColumnOptions,
   type SelectTableOptions,
   type SelectTableRowOptions,
   type TableCommandsExtension,
 } from './table-commands'
+export {
+  moveTableColumn,
+  type MoveTableColumnOptions,
+} from './table-commands/move-table-column'
+export {
+  moveTableRow,
+  type MoveTableRowOptions,
+} from './table-commands/move-table-row'
 export { defineTablePlugins } from './table-plugins'
 export {
   defineTableCellSpec,
