+++ packages/extensions/src/table/table-commands.ts
@@ -27,11 +27,17 @@ import {
 } from 'prosemirror-tables'
 
 import {
+  moveTableColumn,
+  type MoveTableColumnOptions,
+} from './table-commands/move-table-column'
+import {
+  moveTableRow,
+  type MoveTableRowOptions,
+} from './table-commands/move-table-row'
+import {
   findCellPos,
   findCellRange,
   findTable,
-  moveColumn,
-  moveRow,
 } from './table-utils'
 
 function createEmptyTable(
@@ -293,106 +299,6 @@ export function selectTable(options?: SelectTableOptions): Command {
       tr = tr.setSelection(selection)
       dispatch?.(tr)
     }
-    return true
-  }
-}
-
-/**
- * Options for {@link moveTableRow}
- *
- * @public
- */
-export interface MoveTableRowOptions {
-  /**
-   * The source row index to move from.
-   */
-  origin: number
-
-  /**
-   * The destination row index to move to.
-   */
-  target: number
-
-  /**
-   * Whether to select the moved row after the operation.
-   *
-   * @default true
-   */
-  select?: boolean
-
-  /**
-   * Optional position to resolve table from. If not provided, uses the current selection.
-   */
-  pos?: number
-}
-
-/**
- * Move a table row from index `origin` to index `target`.
- *
- * @public
- */
-export function moveTableRow(options: MoveTableRowOptions): Command {
-  return (state, dispatch) => {
-    const { origin, target, select = true, pos } = options
-    const tr = moveRow({
-      tr: state.tr,
-      origin,
-      target,
-      select,
-      pos,
-    })
-
-    dispatch?.(tr)
-    return true
-  }
-}
-
-/**
- * Options for {@link moveTableColumn}
- *
- * @public
- */
-export interface MoveTableColumnOptions {
-  /**
-   * The source column index to move from.
-   */
-  origin: number
-
-  /**
-   * The destination column index to move to.
-   */
-  target: number
-
-  /**
-   * Whether to select the moved column after the operation.
-   *
-   * @default true
-   */
-  select?: boolean
-
-  /**
-   * Optional position to resolve table from. If not provided, uses the current selection.
-   */
-  pos?: number
-}
-
-/**
- * Move a table column from index `origin` to index `target`.
- *
- * @public
- */
-export function moveTableColumn(options: MoveTableColumnOptions): Command {
-  return (state, dispatch) => {
-    const { origin, target, select = true, pos } = options
-    const tr = moveColumn({
-      tr: state.tr,
-      origin,
-      target,
-      select,
-      pos,
-    })
-
-    dispatch?.(tr)
     return true
   }
 }
