+++ packages/extensions/src/table/table-utils/convert-array-of-rows-to-table-node.spec.ts
