+++ packages/extensions/src/table/table-utils/convert-table-node-to-array-of-rows.spec.ts
