+++ packages/extensions/src/table/table-utils/move-row.ts
@@ -7,8 +7,8 @@ import {
 
 import { convertArrayOfRowsToTableNode } from './convert-array-of-rows-to-table-node'
 import { convertTableNodeToArrayOfRows } from './convert-table-node-to-array-of-rows'
-import { moveRowInArrayOfRows } from './move-row-in-array-of-rows'
 import { getSelectionRangeInRow } from './get-selection-range-in-row'
+import { moveRowInArrayOfRows } from './move-row-in-array-of-rows'
 import { findTable } from './query'
 
 export interface MoveRowParams {
